+++ packages/application-extension/src/index.tsx
@@ -18,8 +18,6 @@ import {
   IDocumentProvider,
   IDocumentProviderFactory,
   ProviderMock,
-  getAnonymousUserName,
-  getRandomColor,
 } from '@jupyterlab/docprovider';
 
 import { IFileBrowserFactory } from '@jupyterlab/filebrowser';
@@ -36,15 +34,9 @@ import { liteIcon, liteWordmark } from '@jupyterlite/ui-components';
 
 import { filter, toArray } from '@lumino/algorithm';
 
-import { UUID, PromiseDelegate } from '@lumino/coreutils';
-
 import { Widget } from '@lumino/widgets';
 
 import { getParam } from 'lib0/environment';
-
-import { WebrtcProvider } from 'y-webrtc';
-
-import { Awareness } from 'y-protocols/awareness';
 
 import React from 'react';
 
@@ -63,113 +55,6 @@ const EDITOR_FACTORY = 'Editor';
  */
 const URL_PATTERN = new RegExp('/(lab|notebooks|edit|consoles)\\/?');
 
-class WebRtcProvider extends WebrtcProvider implements IDocumentProvider {
-  constructor(options: IWebRtcProvider.IOptions) {
-    super(
-      `${options.room}${options.path}`,
-      options.ymodel.ydoc,
-      WebRtcProvider.yProviderOptions(options)
-    );
-    this.awareness = options.ymodel.awareness;
-    const color = `#${getParam('--usercolor', getRandomColor().slice(1))}`;
-    const name = getParam('--username', getAnonymousUserName());
-    const currState = this.awareness.getLocalState();
-    // only set if this was not already set by another plugin
-    if (currState && !currState.name) {
-      this.awareness.setLocalStateField('user', {
-        name,
-        color,
-      });
-    }
-  }
-
-  setPath() {
-    // TODO: this seems super useful
-  }
-
-  requestInitialContent(): Promise<boolean> {
-    if (this._initialRequest) {
-      return this._initialRequest.promise;
-    }
-    let resolved = false;
-    this._initialRequest = new PromiseDelegate<boolean>();
-    this.on('synced', (event: any) => {
-      if (this._initialRequest) {
-        this._initialRequest.resolve(event.synced);
-        resolved = true;
-      }
-    });
-    // similar logic as in the upstream plugin
-    setTimeout(() => {
-      if (!resolved && this._initialRequest) {
-        this._initialRequest.resolve(false);
-      }
-    }, 1000);
-    return this._initialRequest.promise;
-  }
-
-  putInitializedState(): void {
-    // no-op
-  }
-
-  acquireLock(): Promise<number> {
-    return Promise.resolve(0);
-  }
-
-  releaseLock(lock: number): void {
-    // no-op
-  }
-
-  private _initialRequest: PromiseDelegate<boolean> | null = null;
-}
-
-/**
- * A public namespace for WebRTC options
- */
-export namespace IWebRtcProvider {
-  export interface IOptions extends IDocumentProviderFactory.IOptions {
-    room: string;
-    signalingUrls?: string[];
-  }
-
-  export interface IYjsWebRtcOptions {
-    signaling: Array<string>;
-    password: string | null;
-    awareness: Awareness;
-    maxConns: number;
-    filterBcConns: boolean;
-    peerOpts: any;
-  }
-}
-
-/**
- * A private (so far) namespace for Yjs/WebRTC implementation details
- */
-namespace WebRtcProvider {
-  /**
-   * Re-map Lab provider options to yjs ones.
-   */
-  export function yProviderOptions(
-    options: IWebRtcProvider.IOptions
-  ): IWebRtcProvider.IYjsWebRtcOptions {
-    return {
-      signaling:
-        options.signalingUrls && options.signalingUrls.length
-          ? options.signalingUrls
-          : [
-              'wss://signaling.yjs.dev',
-              'wss://y-webrtc-signaling-eu.herokuapp.com',
-              'wss://y-webrtc-signaling-us.herokuapp.com',
-            ],
-      password: null,
-      awareness: new Awareness(options.ymodel.ydoc),
-      maxConns: 20 + Math.floor(Math.random() * 15), // the random factor reduces the chance that n clients form a cluster
-      filterBcConns: true,
-      peerOpts: {}, // simple-peer options. See https://github.com/feross/simple-peer#peer--new-peeropts
-    };
-  }
-}
-
 /**
  * The command IDs used by the application extension.
  */
@@ -182,6 +67,12 @@ namespace CommandIDs {
 
   export const copyShareableLink = 'filebrowser:share-main';
 }
+
+/**
+ * The name of the translation bundle for internationalized strings.
+ */
+
+const I18N_BUNDLE = 'jupyterlite';
 
 /**
  * Add a command to show an About dialog.
@@ -198,7 +89,7 @@ const about: JupyterFrontEndPlugin<void> = {
     menu: IMainMenu | null
   ): void => {
     const { commands } = app;
-    const trans = translator.load('jupyterlab');
+    const trans = translator.load(I18N_BUNDLE);
     const category = trans.__('Help');
 
     commands.addCommand(CommandIDs.about, {
@@ -245,7 +136,7 @@ const about: JupyterFrontEndPlugin<void> = {
         );
         const copyright = (
           <span className="jp-About-copyright">
-            {trans.__('© 2021 JupyterLite Contributors')}
+            {trans.__('© 2021-2022 JupyterLite Contributors')}
           </span>
         );
         const body = (
@@ -284,24 +175,27 @@ const about: JupyterFrontEndPlugin<void> = {
 const docProviderPlugin: JupyterFrontEndPlugin<IDocumentProviderFactory> = {
   id: '@jupyterlite/application-extension:docprovider',
   provides: IDocumentProviderFactory,
-  activate: (app: JupyterFrontEnd): IDocumentProviderFactory => {
-    const roomName = getParam('--room', '').trim();
-    const host = window.location.host;
-    // enable if both the page config option (deployment wide) and the room name (user) are defined
-    const collaborative = PageConfig.getOption('collaborative') === 'true' && roomName;
-    const signalingUrls = JSON.parse(
-      PageConfig.getOption('fullWebRtcSignalingUrls') || 'null'
-    );
-    // default to a random id to not collaborate with others by default
-    const room = `${host}-${roomName || UUID.uuid4()}`;
+  requires: [ITranslator],
+  activate: (
+    app: JupyterFrontEnd,
+    translator: ITranslator
+  ): IDocumentProviderFactory => {
+    const collaborative = PageConfig.getOption('collaborative') === 'true';
     const factory = (options: IDocumentProviderFactory.IOptions): IDocumentProvider => {
-      return collaborative
-        ? new WebRtcProvider({
-            room,
-            ...options,
-            ...(signalingUrls && signalingUrls.length ? { signalingUrls } : {}),
-          })
-        : new ProviderMock();
+      if (collaborative) {
+        const trans = translator.load(I18N_BUNDLE);
+        console.warn(
+          trans.__(
+            'The `collaborative` feature was enabled, but no docprovider is available.'
+          )
+        );
+        console.info(
+          trans.__(
+            'Install `jupyterlab-webrtc-docprovider` to enable WebRTC-based collaboration.'
+          )
+        );
+      }
+      return new ProviderMock();
     };
     return factory;
   },
@@ -322,7 +216,7 @@ const downloadPlugin: JupyterFrontEndPlugin<void> = {
     palette: ICommandPalette | null,
     factory: IFileBrowserFactory | null
   ) => {
-    const trans = translator.load('jupyterlab');
+    const trans = translator.load(I18N_BUNDLE);
     const { commands, serviceManager, shell } = app;
     const { contents } = serviceManager;
 
@@ -366,6 +260,7 @@ const downloadPlugin: JupyterFrontEndPlugin<void> = {
             buttons: [Dialog.okButton({ label: trans.__('OK') })],
           });
         }
+        await context.save();
         const content = await formatContent(context.path);
         downloadContent(content, context.path);
       },
@@ -452,10 +347,12 @@ const opener: JupyterFrontEndPlugin<void> = {
   id: '@jupyterlite/application-extension:opener',
   autoStart: true,
   requires: [IRouter, IDocumentManager],
+  optional: [ILabShell],
   activate: (
     app: JupyterFrontEnd,
     router: IRouter,
-    docManager: IDocumentManager
+    docManager: IDocumentManager,
+    labShell: ILabShell | null
   ): void => {
     const { commands } = app;
 
@@ -504,6 +401,16 @@ const opener: JupyterFrontEndPlugin<void> = {
               url.searchParams.delete('path');
               const { pathname, search } = url;
               router.navigate(`${pathname}${search}`, { skipRouting: true });
+
+              if (labShell) {
+                // open the folder where the files are located on startup
+                const showInBrowser = () => {
+                  commands.execute('docmanager:show-in-file-browser');
+                  labShell.currentChanged.disconnect(showInBrowser);
+                };
+
+                labShell.currentChanged.connect(showInBrowser);
+              }
               break;
             }
           }
@@ -531,7 +438,7 @@ const shareFile: JupyterFrontEndPlugin<void> = {
     factory: IFileBrowserFactory,
     translator: ITranslator
   ): void => {
-    const trans = translator.load('jupyterlab');
+    const trans = translator.load(I18N_BUNDLE);
     const { commands } = app;
     const { tracker } = factory;