+++ packages/kernel/src/tokens.ts
@@ -1,3 +1,8 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+
+import type { Remote } from 'comlink';
+
 import { Kernel, KernelMessage, KernelSpec } from '@jupyterlab/services';
 
 import { Token } from '@lumino/coreutils';
@@ -12,6 +17,11 @@ import { KernelSpecs } from './kernelspecs';
  * The token for the kernels service.
  */
 export const IKernels = new Token<IKernels>('@jupyterlite/kernel:IKernels');
+
+/**
+ * The kernel name of last resort.
+ */
+export const FALLBACK_KERNEL = 'javascript';
 
 /**
  * An interface for the Kernels service.
@@ -54,6 +64,11 @@ export interface IKernel extends IObservableDisposable {
   readonly name: string;
 
   /**
+   * The location in the virtual filesystem from which the kernel was started.
+   */
+  readonly location: string;
+
+  /**
    * A promise that is fulfilled when the kernel is ready.
    */
   readonly ready: Promise<void>;
@@ -90,6 +105,11 @@ export namespace IKernel {
     name: string;
 
     /**
+     * The location in the virtual filesystem from which the kernel was started.
+     */
+    location: string;
+
+    /**
      * The method to send messages back to the server.
      */
     sendMessage: SendMessage;
@@ -111,6 +131,11 @@ export interface IKernelSpecs {
   readonly specs: KernelSpec.ISpecModels | null;
 
   /**
+   * Get the default kernel name.
+   */
+  readonly defaultKernelName: string;
+
+  /**
    * Get the kernel factories for the current kernels.
    */
   readonly factories: KernelSpecs.KernelFactories;
@@ -122,3 +147,54 @@ export interface IKernelSpecs {
    */
   register: (options: KernelSpecs.IKernelOptions) => void;
 }
+
+/**
+ * An interface for a comlink-based worker kernel
+ */
+export interface IWorkerKernel {
+  /**
+   * Handle any lazy setup activities.
+   */
+  initialize(options: IWorkerKernel.IOptions): Promise<void>;
+  execute(
+    content: KernelMessage.IExecuteRequestMsg['content'],
+    parent: any
+  ): Promise<KernelMessage.IExecuteReplyMsg['content']>;
+  complete(
+    content: KernelMessage.ICompleteRequestMsg['content'],
+    parent: any
+  ): Promise<KernelMessage.ICompleteReplyMsg['content']>;
+  inspect(
+    content: KernelMessage.IInspectRequestMsg['content'],
+    parent: any
+  ): Promise<KernelMessage.IInspectReplyMsg['content']>;
+  isComplete(
+    content: KernelMessage.IIsCompleteRequestMsg['content'],
+    parent: any
+  ): Promise<KernelMessage.IIsCompleteReplyMsg['content']>;
+  commInfo(
+    content: KernelMessage.ICommInfoRequestMsg['content'],
+    parent: any
+  ): Promise<KernelMessage.ICommInfoReplyMsg['content']>;
+  commOpen(content: KernelMessage.ICommOpenMsg, parent: any): Promise<void>;
+  commMsg(content: KernelMessage.ICommMsgMsg, parent: any): Promise<void>;
+  commClose(content: KernelMessage.ICommCloseMsg, parent: any): Promise<void>;
+  inputReply(
+    content: KernelMessage.IInputReplyMsg['content'],
+    parent: any
+  ): Promise<void>;
+}
+
+/**
+ * A namespace for worker kernels.
+ **/
+export namespace IWorkerKernel {
+  /**
+   * Common values likely to be required by all kernels.
+   */
+  export interface IOptions {
+    baseUrl: string;
+  }
+}
+
+export interface IRemoteKernel extends Remote<IWorkerKernel> {}